deploy: remove unused imports and adjust type definitions
