+++ src/service/auth.service.ts
@@ -6,12 +6,68 @@ type Credentials = {
   email: string;
   password: string;
 };
+//MOCK
 
+const clientSession = {
+  user: {
+    id: 6,
+    email: "fazenda@gmail.com",
+    phone: "+244 912 345 678",
+    nif: "986453675",
+    role: "COMMON" as "COMMON",
+    profile_type: "individual" as "individual",
+    photo:
+      "https://blog4.mfrural.com.br/wp-content/uploads/2020/02/fazendas-1024x660.jpg",
+    created_at: "2025-04-11T13:03:20.000Z",
+    isActive: true,
+    twoFactorEnabled: false,
+    first_name: "Fazenda",
+    last_name: "Client",
+  },
+  access_token: "eyJhbGciOiJIUzI...",
+};
+const adminSesison = {
+  user: {
+    id: 6,
+    email: "admin@gmail.com",
+    phone: "+244 912 345 678",
+    nif: "986453675",
+    role: "ADMIN" as "ADMIN",
+    profile_type: "individual" as "individual",
+    photo:
+      "https://blog4.mfrural.com.br/wp-content/uploads/2020/02/fazendas-1024x660.jpg",
+    created_at: "2025-04-11T13:03:20.000Z",
+    isActive: true,
+    twoFactorEnabled: false,
+    first_name: "Fazenda",
+    last_name: "Admin",
+  },
+  access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlb...",
+};
+//---------------
 // service/auth.service.ts
 export async function login(credentials: Credentials): Promise<Session> {
-  const response = await api.post(`/auth/login`, credentials);
-  const { access_token, user } = response.data;
-  return { access_token, user };
+  if (import.meta.env.VITE_INTEGRATION_IN_PROGRESS === "true") {
+    console.log("a");
+    localStorage.setItem("access_token", clientSession.access_token);
+    localStorage.setItem("user", JSON.stringify(clientSession.user));
+    if (credentials.email === "admin@gmail.com") {
+      return {
+        access_token: adminSesison.access_token,
+        user: adminSesison.user,
+      };
+    } else {
+      return {
+        access_token: clientSession.access_token,
+        user: clientSession.user,
+      };
+    }
+  } else {
+    console.log("b");
+    const response = await api.post(`/auth/login`, credentials);
+    const { access_token, user } = response.data;
+    return { access_token, user };
+  }
 }
 
 export async function registerUser(
